Refactor gradio into separate file and isolate it from torch specific stuff
